requirements.txt
absl-py==2.3.1
annotated-types==0.7.0
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.1.8
contourpy==1.1.1
cycler==0.12.1
diffusers==0.10.2
efficientnet_pytorch==0.7.1
eval_type_backport==0.2.2
filelock==3.16.1
fonttools==4.57.0
fsspec==2025.3.0
ftfy==6.2.3
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.40.3
google-auth-oauthlib==1.0.0
grpcio==1.70.0
hf-xet==1.1.9
huggingface-hub==0.12.1
idna==3.10
imageio==2.35.1
importlib_metadata==8.5.0
importlib_resources==6.4.5
Jinja2==3.1.6
joblib==1.4.2
kiwisolver==1.4.7
lazy_loader==0.4
lightning-utilities==0.11.9
Markdown==3.7
MarkupSafe==2.1.5
matplotlib==3.7.5
mpmath==1.3.0
networkx==3.0
numpy==1.24.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu12==12.1.105
oauthlib==3.3.1
packaging==25.0
pandas==2.0.3
pillow==10.4.0
platformdirs==4.3.6
protobuf==5.29.5
psutil==7.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.10.6
pydantic_core==2.27.2
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2025.2
PyWavelets==1.4.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
requests-oauthlib==2.0.0
rsa==4.9.1
safetensors==0.5.3
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.2
sentry-sdk==2.42.1
six==1.17.0
smmap==5.0.2
sympy==1.13.3
tensorboard==2.14.0
tensorboard-data-server==0.7.2
threadpoolctl==3.5.0
tifffile==2023.7.10
timm==1.0.19
tokenizers==0.13.3
torch==2.4.1+cu121
torch-fidelity==0.3.0
torchaudio==2.4.1+cu121
torchmetrics==1.5.2
torchvision==0.19.1+cu121
tqdm==4.67.1
transformers==4.27.0
triton==3.0.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.2.3
wandb==0.22.2
wcwidth==0.2.13
Werkzeug==3.0.6
zipp==3.20.2
